requirements.txt
streamlit==1.12.0
Pillow==9.2.0
streamlit-drawable-canvas==0.9.3